Make sure producer auto configuration runs after BindingServiceConfiguration
The propagation PR on query service was failing because BindingServiceConfiguration (that creates BindingService bean) was running after TestConfiguration
https://github.com/Activiti/activiti-cloud-query-service/pull/372
